Added Genetic Algorithm Visualization Example
Added a Genetic Algorithm visualization example that uses selection, crossover, mutation, and elitism to continuously evolve the population. Tournament Selection and Roulette Wheel Selection options have been implemented for the type of selection to use in the algorithm. There is also an option for the user to build maze-like "walls" for the algorithm to avoid, to further show how the algorithm works.